Create TradinSimulation.py
This is a stock trading simulation I built. It uses pretty simple factors, that is moving mean and stand deviation, variable buy/sell factor, and variable transaction amount. 

It analyzes the past 700 days and implements this trading strategy throughout those days.

It uses Yahoo Fiance's API, YFinance. 

It's a pretty cool model. But definitely doesn't translate one-to-one in the real market, and is not factoring in many things.