Create portfolio_trading_simulation,py
This is a new algorithm that I made. It is similar to the "trading simulation" that I have except this one is trading within a whole portfolio throughout about 700 days (probably a bit less). This allows for more realistic trading as watching one stock for a whole year is not likely a good strategy. However, with a portfolio, we can witness much more transitions.