Create Trend Breakout indicator.py
This code uses historic FOREX candle data. It checks the current trend to try and find a breakout point. It identifies the points that start the trends, and then the points at which it breaks those trends. It uses regression models to create the trend of best fit given certain criteria.