Create break out and support graphics.py
This code provides an algorithm that will identify breakouts and support lines for a price graph. But by using it for history, we can make it look a bit better. We do this by sorting our lines and removing duplicates. Real-time data may make that a bit more difficult. but everything would work the same, except the graphics would be different.