Create moving support and resistance
This code provides the functions I used to create moving support and resistance charts. I created this because looking at the previous ones I built, it was terrible. The old and traditional ones didn't have anything to do with time. This one keeps the resistance and support levels until they change