Create backtesting.py
This one took a while to figure out. All of the resources online that were found did not use pandas TA. Which is what I am most familiar with, so it took a while to be able to incorporate that package. But not, this code trades solely based on RSI, with finding the best fit of the variables RSI upper, RSI lower, and RSI period. The max return was about 600% as a buy and hold was about 700%. Not sure about the full strategy that was implemented. But, during time of excess cash, we could probably invest in something safe to make up for the buy-and-hold opportunity. That is, even though we had a lower percent return, we still might of had greater opportunity to invest into other things